% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{aci}
\alias{aci}
\title{Acoustic Complexity Index}
\usage{
aci(
  wave,
  freq.res = 50,
  w.fun = "hanning",
  min.freq = NA,
  max.freq = NA,
  j = NA,
  noise.red = 0,
  rm.offset = TRUE
)
}
\arguments{
\item{wave}{an object of class Wave imported with the \emph{readWave} function of the \emph{tuneR} package.}

\item{freq.res}{numeric. The frequency resolution to use (Hz per bin) which will determine the window length for the FFT (sampling rate / frequency resolution).}

\item{w.fun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min.freq}{minimum frequency to use when calculating the value, in Hertz. Default = 0.}

\item{max.freq}{maximum frequency to use when calculating the value, in Hertz. Default = NA (Nyquist).}

\item{j}{the cluster size, in seconds. Default = NA (Duration of the audio file).}

\item{noise.red}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0, noise reduction is not applied.}

\item{rm.offset}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}
}
\value{
A tibble (data frame) with the ACI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Acoustic Complexity Index (ACI) from Pieretti, et al. 2011. The ACI is based
on the "observation that many biotic sounds, such as bird songs,
are characterized by an intrinsic variability of intensities,
while some types of human generated noise (such as car passing or airplane transit)
present very constant intensity values" (Pieretti, et al. 2011).
}
\details{
The index was tested to the ACItot calculated using SoundscapeMeter v 1.0.14.05.2012,
courtesy of A. Farina. The results are accumulative.
Very long samples will return comparatively larger values for ACI.
The current version (\emph{soundecology2}) normalizes the output (i.e., "j" equals the
duration of the audio file) to make it equivalent to the default results in seewave's version.

Reference: N. Pieretti, A. Farina, D. Morri. 2011. A new methodology to infer
the singing activity of an avian community: The Acoustic Complexity Index (ACI). Ecological Indicators 11: 868-873.
}
\examples{
\dontrun{
aci(tropicalsound)}
}
