% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci_list.R
\name{aci_list}
\alias{aci_list}
\title{Acoustic Complexity Index - list input}
\usage{
aci_list(
  audiolist,
  save_csv = FALSE,
  csv_name = "aci_results.csv",
  wlen = 512,
  wfun = "hanning",
  min_freq = NA,
  max_freq = NA,
  j = NA,
  noisered = 2,
  rmoff = TRUE
)
}
\arguments{
\item{audiolist}{a list of audio files to import.}

\item{save_csv}{logical. Whether to save a csv in the working directory.}

\item{csv_name}{character vector. When 'save_csv' is TRUE, optionally provide a file name.}

\item{wlen}{the window length to compute the spectrogram (i.e., FFT window size).}

\item{wfun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min_freq}{minimum frequency to use when calculating the value, in Hertz. Default = 0.}

\item{max_freq}{maximum frequency to use when calculating the value, in Hertz. Default = NA (Nyquist).}

\item{j}{the cluster size, in seconds. Default = NA (Duration of the audio file).}

\item{noisered}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0, noise reduction is not applied.}

\item{rmoff}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}
}
\value{
A tibble (data frame) with the ACI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Acoustic Complexity Index - list input
}
\details{
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a folder of audio files before importing them into R.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
files <- list.files(pattern=".wav|.WAV")
aci_list(files[1:5])
}
