% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adi_heatmap.R
\name{adi_heatmap}
\alias{adi_heatmap}
\title{ADI Heatmap}
\usage{
adi_heatmap(df, plot.title)
}
\arguments{
\item{df}{A data frame with ADI data for 1 day}

\item{plot.title}{Character; a title for the plot}
}
\value{
A heatmap plot created with ggplot2.
}
\description{
Produces a heatmap with the proportions used to calculate ADI for a day of data. The current version supports only data frames describing stereo files with a duty cycle of 1 on; 9 off.
}
\examples{
adi_heatap(adi20230903, "ADI proportions")
}
