% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aei_list.R
\name{aei_list}
\alias{aei_list}
\title{Acoustic Evenness Index - list input}
\usage{
aei_list(
  audiolist,
  save_csv = FALSE,
  csv_name = "aei_results.csv",
  wlen = 512,
  wfun = "hanning",
  min_freq = 0,
  max_freq = 10000,
  nbands = 10,
  db_threshold = 5,
  normspec = FALSE,
  noisered = 2,
  rmoff = TRUE,
  props = TRUE,
  propden = 1
)
}
\arguments{
\item{audiolist}{a list with the audio files to import.}

\item{save_csv}{logical. Whether to save a csv in the working directory.}

\item{csv_name}{character vector. When 'save_csv' is TRUE, optionally provide a file name.}

\item{wlen}{window length to compute the spectrogram}

\item{wfun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min_freq}{minimum frequency to compute the spectrogram}

\item{max_freq}{maximum frequency to compute the spectrogram}

\item{nbands}{number of bands to split the spectrogram}

\item{db_threshold}{dB threshold to calculate energy proportions (if normspec = FALSE, set to 5 or above)}

\item{normspec}{logical. Whether to normalize the spectrogram (not recommended) or not (normalized spectrograms with different SNR are not comparable).}

\item{noisered}{numeric. noise reduction (subtract median from the amplitude values); 1=rows, 2=columns, 3=none.}

\item{rmoff}{logical. Whether to remove DC offset before computing aei (recommended) or not.}

\item{props}{logical. Whether to store the energy proportion values for each frequency band and channel (default) or not.}

\item{propden}{numeric. Indicates how the energy proportion is calculated by manipulating the denominator.}
}
\value{
a tibble (data frame) with the aei values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Calculates the Acoustic Evenness Index for all the files in a list, with extended parameter options.
It uses parallel processing with all but one of the available cores.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com)  April 2024.
}
\details{
Options for propden:
1 = The original calculation from the "soundecology" package is applied. The denominator of the proportion equals to all the cells in the same frequency band.
2 = The "whole population across species" equals the cells above the decibel threshold across the spectrogram (up to 'max_freq')
3 = The "whole population across species" equals the cells above the decibel threshold across the whole spectrogram (up to the Nyquist frequency. This might return a smaller range of values.
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a list of audio files before importing them into R.
The working directory should be set to the folder containing the files.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
files <- list.files(pattern=".wav|.WAV")
aei_list(files[1:5])
}
