% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_folder.R
\name{bi_folder}
\alias{bi_folder}
\title{Bioacoustic Index - folder input}
\usage{
bi_folder(
  folder,
  save_csv = FALSE,
  csv_name = "bi_results.csv",
  wlen = 512,
  wfun = "hanning",
  min_freq = 2000,
  max_freq = 8000,
  normspec = FALSE,
  noisered = 0,
  rmoff = TRUE
)
}
\arguments{
\item{folder}{a path to the folder with audio files to import.}

\item{save_csv}{logical. Whether to save a csv in the working directory.}

\item{csv_name}{character vector. When 'save_csv' is TRUE, optionally provide a file name.}

\item{wlen}{the window length to compute the spectrogram (i.e., FFT window size).}

\item{wfun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min_freq}{miminum frequency to use when calculating the value, in Hertz. Default = NA.}

\item{max_freq}{maximum frequency to use when calculating the value, in Hertz. Default = NA (Nyquist).}

\item{normspec}{logical; if TRUE, the spectrogram is normalized, scaled by its maximum value (not recommended because normalized spectrograms with different SNR are not comparable).}

\item{noisered}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0 (Default), noise reduction is not applied.}

\item{rmoff}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}
}
\value{
A tibble (data frame) with the BI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Inspired by the "Bioacoustic Index" from the paper:Boelman NT, Asner GP, Hart PJ, Martin RE. 2007.
Multi-trophic invasion resistance in Hawaii: bioacoustics, field surveys, and airborne
remote sensing. Ecol Applications 17(8):2137-44. Based on Matlab code provided by NT Boelman.
Boelman et al. 2007 used min_freq=2000, max_freq=8000, wlen=512.
Several parts where changed, in particular log math, so this won't be
directly comparable to the original code in the paper.
}
\details{
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a folder of audio files before importing them into R.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
bi_folder(path/to/folder)
}
