% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_list.R
\name{bi_list}
\alias{bi_list}
\title{Bioacoustic Index - list input}
\usage{
bi_list(
  audio.list,
  folder = NULL,
  save.csv = TRUE,
  csv.name = "bi_results.csv",
  w.len = 512,
  w.fun = "hanning",
  min.freq = 2000,
  max.freq = 8000,
  norm.spec = FALSE,
  noise.red = 0,
  rm.offset = TRUE,
  n.cores = -1
)
}
\arguments{
\item{audio.list}{a list of audio files to import.}

\item{folder}{a path to the folder with audio files to import.}

\item{save.csv}{logical. Whether to save a csv in the working directory.}

\item{csv.name}{character vector. When 'save.csv' is TRUE, optionally provide a file name.}

\item{w.len}{the window length to compute the spectrogram (i.e., FFT window size).}

\item{w.fun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min.freq}{miminum frequency to use when calculating BI, in Hertz. Default = NA.}

\item{max.freq}{maximum frequency to use when calculating BI, in Hertz. Default = NA (Nyquist).}

\item{norm.spec}{logical; if TRUE, the spectrogram is normalized, scaled by its maximum value (not recommended because normalized spectrograms with different SNR are not comparable).}

\item{noise.red}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0 (Default), noise reduction is not applied.}

\item{rm.offset}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}

\item{n.cores}{The number of cores to use for parallel processing. Use \code{n.cores = -1} to use all but one core. Default is NULL (single-core processing).}
}
\value{
A tibble (data frame) with the BI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Inspired by the "Bioacoustic Index" from the paper:Boelman NT, Asner GP, Hart PJ, Martin RE. 2007.
Multi-trophic invasion resistance in Hawaii: bioacoustics, field surveys, and airborne
remote sensing. Ecol Applications 17(8):2137-44. Based on Matlab code provided by NT Boelman.
Boelman et al. 2007 used min.freq=2000, max.freq=8000, w.len=512.
Several parts where changed, in particular log math, so this won't be
directly comparable to the original code in the paper.
}
\details{
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a list of audio files before importing them in
The working directory should be set to the folder containing the files.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
files <- list.files(pattern=".wav|.WAV")
bi_list(files[1:5])
}
