% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel_spectrogram.R
\name{diel_spectrogram}
\alias{diel_spectrogram}
\title{Generate a Diel Spectrogram from multiple Audio Recordings}
\usage{
diel_spectrogram(
  folder = NULL,
  recursive = TRUE,
  list = NULL,
  title = "",
  freq_res = 50,
  n.cores = -1,
  cutoff = -100,
  max_amp = -10,
  plot = TRUE,
  save_plot = TRUE,
  save_to = "./diel_plots",
  dc_on = NULL,
  dc_off = NULL,
  target_dates = NULL
)
}
\arguments{
\item{folder}{Path to the folder containing WAV files (default: current
working directory).}

\item{recursive}{Logical. Whether to search files in subfolders. Default: TRUE.}

\item{list}{A list with WAV files to parse. The waves must be in the working
directory (default: NULL).}

\item{title}{Character string. Plot title (default: empty).}

\item{freq_res}{Frequency resolution in Hz for FFT (default: 50)}

\item{n.cores}{Number of cores for parallel processing (-1 for all available
cores)}

\item{cutoff}{Minimum dBFS value (noise floor). Values below this will be
clipped (default: -100)}

\item{max_amp}{Maximum dBFS value to be represented by the color scale.}

\item{plot}{Whether to generate the plot (default: TRUE)}

\item{save_plot}{Logical. Whether to save the plot}

\item{save_to}{Character. If save_plot = TRUE, provide a path to the folder
where the file should be stored. If no path is provided, a new folder
"diel_plots" is created in the current working directory (default) to save
the exported files.}

\item{dc_on}{Number of minutes where the recorded was "on". Used to identify
expected number of files per day.}

\item{dc_off}{Number of minutes where the recorded was "off". Used to identify
expected number of files per day.}

\item{target_dates}{Either a single date or multiple dates in the format
"YYYY-MM-DD" to be analyzed. If NULL (default), all the dates in the folder
will be parsed. To provide a range, see Details.}
}
\value{
A list containing:
\itemize{
\item plot - The ggplot2 spectrogram object
\item spectral_data - Wide-format spectral data (freq bins × recordings)
\item plot_data - Long-format data for plotting
\item file_metadata - File information with extracted metadata
\item overall_range - The dBFS range of all data
}
}
\description{
Creates a spectrogram visualization of daily acoustic patterns using
multiple recordings, with options for noise floor clipping and parallel
processing. Specifically, it takes the variance of acoustic energy in each
frequency bin for each recording and merges the resulting vectors into a
summary spectrogram.
}
\details{
This function parses files in a folder or list and groups them by date,
creating a single diel spectrogram for each. To parse a subset range of dates
contained in a folder, use seq().
Example: target_dates = seq(as.Date("2022-10-20"), as.Date("2022-10-25), "days")
}
