% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_recording_hour.R
\name{first_recording_hour}
\alias{first_recording_hour}
\title{Subsample a dataset with only the first recording in each hour}
\usage{
first_recording_hour(data)
}
\arguments{
\item{data}{a data frame or tibble}
}
\value{
a data frame with the subsetted observations (1 per hour).
}
\description{
Subsample a dataset with only the first recording in each hour
}
\examples{
first_recording_hour(dataframe)
}
