% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kaleidoscope.R
\name{import_kaleidoscope}
\alias{import_kaleidoscope}
\title{Import tidy dataset from Kaleidoscope}
\usage{
import_kaleidoscope(file)
}
\arguments{
\item{data}{An excelfile}
}
\value{
a data frame with the following columns 'file_name', 'sensor_id',
'datetime', 'date', 'hour', 'value_l', 'value_r', 'value_avg'
}
\description{
Import tidy dataset from Kaleidoscope
}
\examples{
import_kaleidoscope(data)
}
