% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stereo_mix.R
\name{import_stereo_mix}
\alias{import_stereo_mix}
\title{Import a Stereo Wave File as Mono and Optionally Remove DC Offset}
\usage{
import_stereo_mix(audio.file, remove.dc.off = TRUE)
}
\arguments{
\item{audio.file}{A string specifying the path to the wave file.}

\item{remove.dc.off}{A logical parameter; if TRUE, the DC offset is removed from the wave object.}
}
\value{
Returns a wave object processed based on the input parameters.
}
\description{
This function reads a stereo wave file, converts it to mono by averaging both channels,
and optionally removes the DC offset from the audio signal. The removal of the DC offset
is controlled by the \code{remove.dc.off} parameter. This function uses the \pkg{tuneR}
package to read and convert the wave file and the \pkg{seewave} package to remove the DC offset.
}
\examples{
# Import a wave file as mono with DC offset removal
wave <- import.as.mono("path/to/your/audiofile.wav")

}
