% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_waves.R
\name{list_waves}
\alias{list_waves}
\title{List wave files in a directory}
\usage{
list_waves(folder = NULL, recursive = FALSE)
}
\arguments{
\item{folder}{Character. The path to the folder where the WAV files are stored.}

\item{recursive}{Logical. Whether to search in subfolders. Default is FALSE.}
}
\value{
A list of wave files
}
\description{
List wave files in a directory
}
\examples{
\dontrun{
list_waves(folder)
}
}
