% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_spectrum_flip.R
\name{mean_spectrum_flip}
\alias{mean_spectrum_flip}
\title{Plot a Mean Spectrum with Flipped Axes}
\usage{
mean_spectrum_flip(
  wave,
  freq.res = 50,
  normalize = FALSE,
  PSD = FALSE,
  PMF = FALSE,
  plot.title = "",
  verbose = TRUE
)
}
\arguments{
\item{wave}{A wave object from the \code{tuneR} or \code{seewave} package, representing the audio signal to be analyzed.}

\item{freq.res}{A numeric value representing the desired frequency resolution. This controls the window length (\code{wl}) used in the \code{meanspec} function. Default is 50.}

\item{normalize}{Logical. Should the spectrum be normalized (i.e., expressed in decibels)? Passed to the \code{norm} argument of \code{meanspec}. Default is \code{FALSE}.}

\item{PSD}{Logical. If \code{TRUE}, compute the power spectral density (PSD). Default is \code{FALSE}.}

\item{PMF}{Logical. If \code{TRUE}, compute the probability mass function (PMF) of the spectrum. Default is \code{FALSE}.}

\item{plot.title}{A string for the plot's title. Default is an empty string (\code{""}).}

\item{verbose}{Logical. If \code{TRUE}, print information about the window length and the number of frequency bins. Default is \code{TRUE}.}
}
\value{
The function returns a plot of the mean spectrum with flipped axes (amplitude on the x-axis and frequency on the y-axis) and the area under the curve filled with black.
}
\description{
This function is a wrapper around the \code{meanspec} function from the \code{seewave} package. It calculates
the mean spectrum of a wave object, flips the axes so that amplitude is on the x-axis and frequency
is on the y-axis, and fills the area under the curve with a black color.
}
\details{
The function computes the mean spectrum using the \code{meanspec} function from the \code{seewave} package. It calculates the window length (\code{wl}) based on the provided frequency resolution, ensuring it is an even number. The function then plots the amplitude on the x-axis and the frequency on the y-axis, with a filled black area under the curve.
}
\examples{
\dontrun{
  # Load example sound data
  data(tico)
  # Plot the mean spectrum with flipped axes
  mean_spectrum_flip(tico, freq.res = 100, normalize = TRUE, plot.title = "Flipped Mean Spectrum")
}

}
