% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsi.R
\name{ndsi}
\alias{ndsi}
\title{Normalized Difference Soundscape Index}
\usage{
ndsi(
  wave,
  wlen = 1024,
  anthro_min = 1000,
  anthro_max = 2000,
  bio_min = 2000,
  bio_max = 11000
)
}
\arguments{
\item{wave}{an object of class Wave imported with the \emph{readWave} function of the \emph{tuneR} package.}

\item{wlen}{the window length to compute the spectrogram (i.e., FFT window size).}

\item{anthro_min}{minimum value of the range of frequencies of the anthrophony.}

\item{anthro_max}{maximum value of the range of frequencies of the anthrophony.}

\item{bio_min}{minimum value of the range of frequencies of the biophony.}

\item{bio_max}{maximum value of the range of frequencies of the biophony.}
}
\value{
a wide format tibble with NDSI values per channel (if stereo), parameters used and audio metadata
}
\description{
Normalized Difference Soundscape Index (NDSI) from REAL and Kasten, et al. 2012. The NDSI seeks to "estimate the level of anthropogenic disturbance on the soundscape by computing the ratio of human-generated (anthrophony) to biological (biophony) acoustic components found in field collected sound samples" (Kasten, et al. 2012).
}
\examples{
ndsi(tropicalsound)
}
