% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsi_folder.R
\name{ndsi_folder}
\alias{ndsi_folder}
\title{Normalized Difference Soundscape Index - folder input}
\usage{
ndsi_folder(
  folder,
  save_csv = FALSE,
  csv_name = "ndsi_results.csv",
  fft_w = 1024,
  anthro_min = 1000,
  anthro_max = 2000,
  bio_min = 2000,
  bio_max = 11000
)
}
\arguments{
\item{folder}{a path to the folder containing the audio files.}

\item{save_csv}{logical. Whether to save a csv in the working directory.}

\item{csv_name}{character vector. When 'save_csv' is TRUE, optionally provide a file name.}

\item{fft_w}{FFT window size.}

\item{anthro_min}{minimum value of the range of frequencies of the anthrophony.}

\item{anthro_max}{maximum value of the range of frequencies of the anthrophony.}

\item{bio_min}{minimum value of the range of frequencies of the biophony.}

\item{bio_max}{maximum value of the range of frequencies of the biophony.}
}
\value{
a wide format tibble with NDSI values per channel (if stereo), parameters used and audio metadata
}
\description{
Normalized Difference Soundscape Index (NDSI) from REAL and Kasten,
et al. 2012. The NDSI seeks to "estimate the level of anthropogenic
disturbance on the soundscape by computing the ratio of human-generated
(anthrophony) to biological (biophony) acoustic components found in field
collected sound samples" (Kasten, et al. 2012).
This version is optimized to work with a path to the folder containing the audio files.
}
\examples{
path <- readClipboard() #use this to paste the folder path from the clipboard.
ndsi_folder(path)
}
