% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsi_list.R
\name{ndsi_list}
\alias{ndsi_list}
\title{Normalized Difference Soundscape Index - list imput}
\usage{
ndsi_list(
  audiolist,
  save_csv = FALSE,
  csv_name = "ndsi_results.csv",
  fft_w = 1024,
  anthro_min = 1000,
  anthro_max = 2000,
  bio_min = 2000,
  bio_max = 11000
)
}
\arguments{
\item{audiolist}{a list of audio files from the working directory.}

\item{save_csv}{logical. Whether to save a csv in the working directory.}

\item{csv_name}{character vector. When 'save_csv' is TRUE, optionally provide a file name.}

\item{fft_w}{FFT window size.}

\item{anthro_min}{minimum value of the range of frequencies of the anthrophony.}

\item{anthro_max}{maximum value of the range of frequencies of the anthrophony.}

\item{bio_min}{minimum value of the range of frequencies of the biophony.}

\item{bio_max}{maximum value of the range of frequencies of the biophony.}
}
\value{
a wide format tibble with NDSI values per channel (if stereo), parameters used and audio metadata
}
\description{
Normalized Difference Soundscape Index (NDSI) from REAL and Kasten,
et al. 2012. The NDSI seeks to "estimate the level of anthropogenic
disturbance on the soundscape by computing the ratio of human-generated
(anthrophony) to biological (biophony) acoustic components found in field
collected sound samples" (Kasten, et al. 2012). This version is optimized to work with lists of files.
}
\details{
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a list of audio files before importing them in
The working directory should be set to the folder containing the files.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
files <- list.files(pattern=".wav|.WAV")
ndsi_list(files[1:5])
}
