% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.list.R
\name{print.list}
\alias{print.list}
\title{Print a list vertically}
\usage{
\method{print}{list}(list)
}
\arguments{
\item{list}{a list object.}
}
\value{
a printed list.
}
\description{
Print a list vertically
}
\examples{
print.list(list)
}
