% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_list.R
\name{print_list}
\alias{print_list}
\title{Print a list vertically}
\usage{
print_list(list)
}
\arguments{
\item{list}{a list object.}
}
\value{
a printed list.
}
\description{
Print a list vertically
}
\examples{
print_list(list)
}
