% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoundEcology2-package.R
\docType{package}
\name{SoundEcology2-package}
\alias{SoundEcology2}
\alias{SoundEcology2-package}
\title{SoundEcology2: Soundscape Ecology}
\description{
Computes acoustic indices on audio recordings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/panchorivasf/SoundEcology2}
  \item Report bugs at \url{https://github.com/panchorivasf/SoundEcology2/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rivas \email{frivasfu@purdue.edu} (\href{https://orcid.org/0000-0002-1256-2225}{ORCID})

Authors:
\itemize{
  \item Luis Villanueva-Rivera
  \item Bryan Pijanowski \email{bpijanow@purdue.edu}
}

Other contributors:
\itemize{
  \item Angel Ramirez \email{azahramirez@gmail.com} [contributor]
}

}
\keyword{internal}
