% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram_cutoff.R
\name{spectrogram_cutoff}
\alias{spectrogram_cutoff}
\title{Generate a Spectrogram with Energy Cutoff}
\usage{
spectrogram_cutoff(
  wave,
  cutoff = -60,
  freq.res = 100,
  plot = FALSE,
  plot.title = NULL,
  ggplot = TRUE,
  noise.red = NULL
)
}
\arguments{
\item{wave}{A \code{Wave} object containing the audio data.}

\item{cutoff}{A numeric value specifying the dB cutoff threshold for the spectrogram. Values below this threshold will be clipped to \code{NA}.}

\item{plot}{A logical value indicating whether to plot the resulting spectrogram. Defaults to \code{FALSE}.}

\item{plot.title}{An optional string specifying the title of the plot. Defaults to \code{NULL}.}

\item{ggplot}{A logical value indicating whether to use ggplot2 for plotting. If \code{FALSE}, base R plotting will be used to generate the plot. Defaults to \code{TRUE}.}
}
\value{
Returns the spectrogram matrix with the cutoff applied. If \code{plot = TRUE}, a plot of the spectrogram is displayed.
}
\description{
This function generates a spectrogram from a wave object and applies a cutoff to filter out
amplitude values below a specified threshold. The resulting spectrogram can be plotted using either
ggplot2 or base R plotting functions.
}
\examples{
\dontrun{
wave <- seewave::synth(d=1, f=8000, cf=1000)
spectrogram_cutoff(wave, cutoff = -40, plot = TRUE, plot.title = "Filtered Spectrogram")
}
}
