% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai.R
\name{tai}
\alias{tai}
\title{Trill Activity Index}
\usage{
tai(
  wave,
  channel = "each",
  hpf = 0,
  cutoff = -60,
  n.windows = 120,
  freq.res = 100,
  plot = FALSE,
  plot.title = NULL,
  verbose = TRUE
)
}
\arguments{
\item{wave}{A wave object to be analyzed.}

\item{channel}{Channel or channels to be analyzed. Options are "left", "right", "each", and "mix".}

\item{cutoff}{Numeric. The cutoff in decibels for the spectrogram generation.}

\item{n.windows}{Numeric. Number of time windows to divide the signal into for analysis. Default is 60.}

\item{freq.res}{Numeric. Frequency resolution (in Hz) for the spectrogram analysis. Default is 100.}

\item{plot}{Logical. If TRUE, generates a plot of the trill index over time. Default is TRUE.}

\item{plot.title}{Character. The title to be used for the plot. Default is NULL.}

\item{verbose}{Logical. If TRUE, provides detailed output during the function's execution. Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{summary}: A tibble summarizing TAI statistics, including values for low and mid-frequency noise.
\item \code{spectral}: A 1-column matrix of mean trill index values for each frequency bin
}
}
\description{
This function calculates the trill index of an audio wave object by analyzing the frequency modulation pattern over time. It can operate in either binary or continuous modes and provides options for generating visual plots of the trill activity. The function also identifies potential noise issues in the low- and mid-frequency ranges.
}
\examples{
\dontrun{
# Example usage:
wave <- read_audio("example.wav")
trill_results <- tai(wave, channel = "left", binary = TRUE, plot = TRUE)
}
}
