% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_integrity_batch.R
\name{wave_integrity_batch}
\alias{wave_integrity_batch}
\title{Compute file integrity diagnostics for several folders containing WAV files}
\usage{
wave_integrity_batch(parent_folder, n.cores = -1)
}
\arguments{
\item{parent_folder}{Character. The path to the parent folder containing folders with WAV files.}

\item{n.cores}{Numeric. Number of cores to be used in the parallel processing. Defaults to -1, which means all but one core.}
}
\value{
A tibble with summary data on file size, including the last day with complete recordings and days with corrupted files (i.e., smaller than most).
}
\description{
Compute file integrity diagnostics for several folders containing WAV files
}
\examples{
\dontrun{
wave_integrity_batch(parentFolder)
}
}
