% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci_list.R
\name{aci_list}
\alias{aci_list}
\title{Calculate Acoustic Complexity Index on a List of Wave Files}
\usage{
aci_list(
  audio.list,
  save.csv = TRUE,
  csv.name = "aci_results.csv",
  freq.res = 50,
  w.fun = "hanning",
  min.freq = NA,
  max.freq = NA,
  j = NA,
  noise.red = 2,
  rm.offset = TRUE,
  n.cores = -1
)
}
\arguments{
\item{audio.list}{a list of audio files to import.}

\item{save.csv}{logical. Whether to save a csv in the working directory.}

\item{csv.name}{character vector. When 'save.csv' is TRUE, optionally provide a file name.}

\item{freq.res}{numeric. The frequency resolution to use (Hz per bin) which will determine the window length for the FFT (sampling rate / frequency resolution).}

\item{w.fun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min.freq}{minimum frequency to use when calculating the value, in Hertz. Default = 0.}

\item{max.freq}{maximum frequency to use when calculating the value, in Hertz. Default = NA (Nyquist).}

\item{j}{the cluster size, in seconds. Default = NA (Duration of the audio file).}

\item{noise.red}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0, noise reduction is not applied.}

\item{rm.offset}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}
}
\value{
A tibble (data frame) with the ACI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Calculate Acoustic Complexity Index on a List of Wave Files
}
\details{
It uses parallel processing with all but one of the available cores.
Optimized to facilitate working with a folder of audio files before importing them into R.
Modifications by Francisco Rivas (frivasfu@purdue.edu // fcorivasf@gmail.com) April 2024
}
\examples{
files <- list_waves(pathToFolder)
aci_list(files[1:5])
}
