% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aei.R
\name{aei}
\alias{aei}
\title{Acoustic Evenness Index}
\usage{
aei(
  wave,
  wlen = 512,
  wfun = "hanning",
  min_freq = 0,
  max_freq = 10000,
  nbands = 10,
  db_threshold = 5,
  normspec = FALSE,
  noisered = 2,
  rmoff = TRUE,
  props = TRUE,
  propden = 1
)
}
\arguments{
\item{wave}{an object of class Wave imported with the \emph{readWave} function of the \emph{tuneR} package.}

\item{wlen}{the window length to compute the spectrogram (i.e., FFT window size).}

\item{wfun}{window function (filter to handle spectral leakage); "bartlett", "blackman", "flattop", "hamming", "hanning", or "rectangle".}

\item{min_freq}{minimum frequency to compute the spectrogram.}

\item{max_freq}{maximum frequency to compute the spectrogram.}

\item{nbands}{number of bands to split the spectrogram.}

\item{db_threshold}{dB threshold to calculate energy proportions (if normspec = FALSE, set to 5 or above).}

\item{normspec}{logical; if TRUE, the spectrogram is normalized, scaled by its maximum value (not recommended because normalized spectrograms with different SNR are not comparable).}

\item{noisered}{numeric; controls the application of noise reduction. If set to 1, noise reduction is applied to each row by subtracting the median from the amplitude values. If set to 2, noise reduction is applied to each column similarly. If set to 0, noise reduction is not applied.}

\item{rmoff}{logical; if set to TRUE, the function will remove DC offset before computing ADI. Default = TRUE.}

\item{props}{logical; if set to TRUE, the function stores the energy proportion values for each frequency band and channel. Default = TRUE.}

\item{propden}{numeric; indicates how the energy proportion is calculated.}
}
\value{
A tibble (data frame) with the AEI values for each channel (if stereo), metadata, and the parameters used for the calculation.
}
\description{
Acoustic Evenness Index from Villanueva-Rivera \emph{et al.} 2011.
The AEI is calculated by dividing the spectrogram into frequency bands (default 10),
taking the proportion of energy in each band above an energy threshold (default -50 dBFS),
and then calculating the Gini Coefficient from those proportions.
The new version allows the user to choose between different ways to compute
the proportions before calculating the Gini, among other new parameter options (see Details)
}
\examples{
aei(tropicalsound)
}
