% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsi.R
\name{ndsi}
\alias{ndsi}
\title{Normalized Difference Soundscape Index}
\usage{
ndsi(
  wave,
  freq.res = 50,
  anthro.min = 1000,
  anthro.max = 2000,
  bio.min = 2000,
  bio.max = 11000
)
}
\arguments{
\item{wave}{an object of class Wave imported with the \emph{readWave} function of the \emph{tuneR} package.}

\item{freq.res}{numeric. The frequency resolution to use (Hz per bin) which will determine the window length for the FFT (sampling rate / frequency resolution).}

\item{anthro.min}{minimum value of the range of frequencies of the anthrophony.}

\item{anthro.max}{maximum value of the range of frequencies of the anthrophony.}

\item{bio.min}{minimum value of the range of frequencies of the biophony.}

\item{bio.max}{maximum value of the range of frequencies of the biophony.}
}
\value{
a wide format tibble with NDSI values per channel (if stereo), parameters used and audio metadata
}
\description{
Normalized Difference Soundscape Index (NDSI) from REAL and Kasten, et al. 2012. The NDSI seeks to "estimate the level of anthropogenic disturbance on the soundscape by computing the ratio of human-generated (anthrophony) to biological (biophony) acoustic components found in field collected sound samples" (Kasten, et al. 2012).
}
\examples{
ndsi(tropicalsound)
}
