% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsi_folder.R
\name{ndsi_folder}
\alias{ndsi_folder}
\title{Normalized Difference Soundscape Index - folder input}
\usage{
ndsi_folder(
  folder = NULL,
  list = NULL,
  w.len = 512,
  anthro.min = 1000,
  anthro.max = 2000,
  bio.min = 2000,
  bio.max = 11000,
  rm.offset = TRUE,
  output.csv = "ndsi_results.csv",
  n.cores = -1
)
}
\arguments{
\item{folder}{a path to the folder containing the audio files.}

\item{list}{An optional list (subset) of files in the folder to analyze. If provided,
files outside the list will be excluded.}

\item{w.len}{numeric. Window length for the FFT (sampling rate / frequency resolution).}

\item{anthro.min}{minimum value of the range of frequencies of the anthrophony.}

\item{anthro.max}{maximum value of the range of frequencies of the anthrophony.}

\item{bio.min}{minimum value of the range of frequencies of the biophony.}

\item{bio.max}{maximum value of the range of frequencies of the biophony.}

\item{rm.offset}{logical. Whether to remove the DC offset.}

\item{output.csv}{character vector. When 'save.csv' is TRUE, optionally provide a file name.
Default name is "ndsi_results.csv"}

\item{n.cores}{The number of cores to use for parallel processing. Use \code{n.cores = -1} to use all
but one core. Default is NULL (single-core processing).}
}
\value{
a wide format tibble with NDSI values per channel (if stereo), parameters used and audio
metadata
}
\description{
Normalized Difference Soundscape Index (NDSI) from REAL and Kasten,
et al. 2012. The NDSI seeks to "estimate the level of anthropogenic
disturbance on the soundscape by computing the ratio of human-generated
(anthrophony) to biological (biophony) acoustic components found in field
collected sound samples" (Kasten, et al. 2012).
This version is optimized to work with a path to the folder containing the audio files.
}
\examples{
\dontrun{
path <- readClipboard() #use this to paste the folder path from the clipboard.
ndsi_folder(path)
}
}
