% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_integrity.R
\name{wave_integrity}
\alias{wave_integrity}
\title{Check the Integrity of a Wave File Dataset}
\usage{
wave_integrity(
  folder,
  type = "lines",
  ggplot = TRUE,
  plot.file = "wave_integrity_plot.png",
  log.file = "wave_integrity_log.txt",
  n.cores = -1,
  dump.folder = TRUE,
  tail.folder = TRUE
)
}
\arguments{
\item{folder}{Character. The path to the folder containing the wave files.}

\item{type}{Character. Type of plot to use in the geom call when ggplot is TRUE. Options are "lines", "bars", and "points".}

\item{ggplot}{Logical. If TRUE, the plot is produced with ggplot2. If FALSE, base plot is used.}

\item{plot.file}{Character. A name for the output plot. Include the extension in the string.}

\item{log.file}{Character. A name for the log file (text file) output.}

\item{n.cores}{Numeric. Number of cores to use in parallel processing. Defults to NULL (no parallel processing).}

\item{dump.folder}{Logical. If TRUE, a 'dump' folder is created in the working directory and all the DUMP files are moved to that folder.}

\item{tail.folder}{Logical. If TRUE, a 'tail' folder is created in the working directory and all files belonging to dates after the last date with complete recordings are moved to that folder.}
}
\value{
A tibble with summary data on file size, including the last day with complete recordings and days with corrupted files (i.e., smaller than most).
}
\description{
This function parses all the WAVE files in a folder and checks their size, reporting corrupted files.
}
\examples{
integrity <- wave_integrity("pathToFolder", ggplot = FALSE)
}
